+++ common/build.gradle
@@ -3,7 +3,7 @@ architectury {
 }
 
 loom {
-    accessWidenerPath = file("src/main/resources/examplemod.accesswidener")
+    accessWidenerPath = file("src/main/resources/dimasrpg.accesswidener")
 }
 
 dependencies {
