+++ common/src/main/resources/architectury.common.json
@@ -1,3 +1,3 @@
 {
-  "accessWidener": "examplemod.accesswidener"
+  "accessWidener": "dimasrpg.accesswidener"
 }
