+++ common/src/main/resources/assets/dimasrpg/lang/en_us.json
@@ -1,0 +1,3 @@
+{
+  "item.dimasrpg.example_item": "Example Item"
+}
