--- common/src/main/resources/assets/examplemod/lang/en_us.json
@@ -1,3 +1,0 @@
-{
-  "item.examplemod.example_item": "Example Item"
-}
