+++ common/src/main/resources/dimasrpg-common.mixins.json
@@ -1,9 +1,10 @@
 {
   "required": true,
-  "package": "com.dimasrpg.mixin.fabric",
+  "package": "com.dimasrpg.mixin",
   "compatibilityLevel": "JAVA_17",
   "minVersion": "0.8",
   "client": [
+    "MixinTitleScreen"
   ],
   "mixins": [
   ],
