--- common/src/main/resources/examplemod-common.mixins.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "package": "net.examplemod.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "minVersion": "0.8",
-  "client": [
-    "MixinTitleScreen"
-  ],
-  "mixins": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
