+++ fabric-like/src/main/java/com/dimasrpg/fabriclike/DimasRPGFabricLike.java
@@ -1,0 +1,9 @@
+package com.dimasrpg.fabriclike;
+
+import com.dimasrpg.DimasRPG;
+
+public class DimasRPGFabricLike {
+    public static void init() {
+        DimasRPG.init();
+    }
+}
