--- fabric-like/src/main/java/net/examplemod/fabriclike/ExampleModFabricLike.java
@@ -1,9 +1,0 @@
-package net.examplemod.fabriclike;
-
-import net.examplemod.ExampleMod;
-
-public class ExampleModFabricLike {
-    public static void init() {
-        ExampleMod.init();
-    }
-}
