+++ fabric/src/main/java/com/dimasrpg/fabric/ExampleExpectPlatformImpl.java
@@ -1,0 +1,15 @@
+package com.dimasrpg.fabric;
+
+import com.dimasrpg.ExampleExpectPlatform;
+import net.fabricmc.loader.api.FabricLoader;
+
+import java.nio.file.Path;
+
+public class ExampleExpectPlatformImpl {
+    /**
+     * This is our actual method to {@link ExampleExpectPlatform#getConfigDirectory()}.
+     */
+    public static Path getConfigDirectory() {
+        return FabricLoader.getInstance().getConfigDir();
+    }
+}
