--- fabric/src/main/java/net/examplemod/fabric/ExampleExpectPlatformImpl.java
@@ -1,15 +1,0 @@
-package net.examplemod.fabric;
-
-import net.examplemod.ExampleExpectPlatform;
-import net.fabricmc.loader.api.FabricLoader;
-
-import java.nio.file.Path;
-
-public class ExampleExpectPlatformImpl {
-    /**
-     * This is our actual method to {@link ExampleExpectPlatform#getConfigDirectory()}.
-     */
-    public static Path getConfigDirectory() {
-        return FabricLoader.getInstance().getConfigDir();
-    }
-}
