+++ fabric/src/main/resources/dimasrpg.mixins.json
