+++ forge/src/main/java/com/dimasrpg/forge/ExampleExpectPlatformImpl.java
@@ -1,0 +1,15 @@
+package com.dimasrpg.forge;
+
+import com.dimasrpg.ExampleExpectPlatform;
+import net.minecraftforge.fml.loading.FMLPaths;
+
+import java.nio.file.Path;
+
+public class ExampleExpectPlatformImpl {
+    /**
+     * This is our actual method to {@link ExampleExpectPlatform#getConfigDirectory()}.
+     */
+    public static Path getConfigDirectory() {
+        return FMLPaths.CONFIGDIR.get();
+    }
+}
