+++ forge/src/main/resources/META-INF/mods.toml
@@ -4,30 +4,30 @@ loaderVersion = "[47,)"
 license = "Insert License Here"
 
 [[mods]]
-modId = "examplemod"
+modId = "dimasrpg"
 version = "${version}"
-displayName = "Example Mod"
-authors = "Me!"
+displayName = "DimasRPG"
+authors = "DimasStudio"
 description = '''
 This is an example description! Tell everyone what your mod is about!
 '''
 #logoFile = ""
 
-[[dependencies.examplemod]]
+[[dependencies.dimasrpg]]
 modId = "forge"
 mandatory = true
 versionRange = "[47,)"
 ordering = "NONE"
 side = "BOTH"
 
-[[dependencies.examplemod]]
+[[dependencies.dimasrpg]]
 modId = "minecraft"
 mandatory = true
 versionRange = "[1.20.1,)"
 ordering = "NONE"
 side = "BOTH"
 
-[[dependencies.examplemod]]
+[[dependencies.dimasrpg]]
 modId = "architectury"
 mandatory = true
 versionRange = "[9.1.12,)"
