+++ forge/src/main/resources/dimasrpg.mixins.json
@@ -1,6 +1,6 @@
 {
   "required": true,
-  "package": "com.dimasrpg.mixin.fabric",
+  "package": "com.dimasrpg.mixin.forge",
   "compatibilityLevel": "JAVA_17",
   "minVersion": "0.8",
   "client": [
