--- forge/src/main/resources/examplemod.mixins.json
@@ -1,13 +1,0 @@
-{
-  "required": true,
-  "package": "net.examplemod.mixin.forge",
-  "compatibilityLevel": "JAVA_17",
-  "minVersion": "0.8",
-  "client": [
-  ],
-  "mixins": [
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
