+++ gradle.properties
@@ -3,9 +3,9 @@ org.gradle.jvmargs=-Xmx2048M
 minecraft_version=1.20.1
 enabled_platforms=quilt,fabric,forge
 
-archives_base_name=examplemod
-mod_version=1.0.0
-maven_group=net.examplemod
+archives_base_name=dimasrpg
+mod_version=0.0.1
+maven_group=com.dimasrpg
 
 architectury_version=9.1.12
