+++ old3/build.gradle
@@ -39,11 +39,11 @@ minecraft {
 
             property 'forge.logging.console.level', 'debug'
 
-            property 'forge.enabledGameTestNamespaces', 'examplemod'
+            property 'forge.enabledGameTestNamespaces', 'dimasrpg'
 
             arg "-mixin.config=mixins.dimasrpg.json"
             mods {
-                examplemod {
+                dimasrpg {
                     source sourceSets.main
                 }
             }
@@ -58,11 +58,11 @@ minecraft {
 
             property 'forge.logging.console.level', 'debug'
 
-            property 'forge.enabledGameTestNamespaces', 'examplemod'
+            property 'forge.enabledGameTestNamespaces', 'dimasrpg'
 
             arg "-mixin.config=mixins.dimasrpg.json"
             mods {
-                examplemod {
+                dimasrpg {
                     source sourceSets.main
                 }
             }
@@ -77,10 +77,10 @@ minecraft {
             property 'mixin.env.remapRefMap', 'true'
             property 'mixin.env.refMapRemappingFile', "${projectDir}/build/createSrgToMcp/output.srg"
 
-            property 'forge.enabledGameTestNamespaces', 'examplemod'
+            property 'forge.enabledGameTestNamespaces', 'dimasrpg'
 
             mods {
-                examplemod {
+                dimasrpg {
                     source sourceSets.main
                 }
             }
@@ -96,7 +96,7 @@ minecraft {
             args '--tweakClass', 'org.spongepowered.asm.launch.MixinTweaker', '--mixin', 'mixins.dimasrpg.json', '--mod', 'dimasrpg', '--all', '--output', file('src/generated/resources/'), '--existing', file('src/main/resources/')
 
             mods {
-                examplemod {
+                dimasrpg {
                     source sourceSets.main
                 }
             }
@@ -126,8 +126,8 @@ dependencies {
 jar {
     manifest {
         attributes([
-                "Specification-Title"     : "examplemod",
-                "Specification-Vendor"    : "examplemodsareus",
+                "Specification-Title"     : "dimasrpg",
+                "Specification-Vendor"    : "dimasrpgsareus",
                 "Specification-Version"   : "1", // We are version 1 of ourselves
                 "Implementation-Title"    : project.name,
                 "Implementation-Version"  : project.jar.archiveVersion,
@@ -135,7 +135,7 @@ jar {
                 "MixinConfigs": "mixins.dimasrpg.json", // Конфигурации миксинов, через зяпятую если их несколько
                 "FMLCorePluginContainsFMLMod": "true",  // Не забываем указать эти два флага,
                 "ForceLoadAsMod": "true",               // чтобы Forge нормально загрузил и мод, и миксины
-                "Implementation-Vendor"   : "examplemodsareus",
+                "Implementation-Vendor"   : "dimasrpgsareus",
                 "Implementation-Timestamp": new Date().format("yyyy-MM-dd'T'HH:mm:ssZ")
         ])
     }
