+++ old3/src/main/resources/pack.mcmeta
@@ -1,6 +1,6 @@
 {
     "pack": {
-        "description": "examplemod resources",
+        "description": "dimasrpg resources",
         "pack_format": 12,
         "forge:resource_pack_format": 12,
         "forge:data_pack_format": 10
