+++ quilt/src/main/java/com/dimasrpg/fabric/ExampleExpectPlatformImpl.java
@@ -1,0 +1,15 @@
+package com.dimasrpg.fabric;
+
+import com.dimasrpg.ExampleExpectPlatform;
+import org.quiltmc.loader.api.QuiltLoader;
+
+import java.nio.file.Path;
+
+public class ExampleExpectPlatformImpl {
+    /**
+     * This is our actual method to {@link ExampleExpectPlatform#getConfigDirectory()}.
+     */
+    public static Path getConfigDirectory() {
+        return QuiltLoader.getConfigDir();
+    }
+}
