+++ quilt/src/main/java/com/dimasrpg/quilt/DimasRPGQuilt.java
@@ -1,0 +1,12 @@
+package com.dimasrpg.quilt;
+
+import com.dimasrpg.fabriclike.DimasRPGFabricLike;
+import org.quiltmc.loader.api.ModContainer;
+import org.quiltmc.qsl.base.api.entrypoint.ModInitializer;
+
+public class DimasRPGQuilt implements ModInitializer {
+    @Override
+    public void onInitialize(ModContainer mod) {
+        DimasRPGFabricLike.init();
+    }
+}
